+++ dask_ml/metrics/regression.py
@@ -1,5 +1,6 @@
 import six
 
+import dask.array as da
 from dask.array.random import doc_wraps
 import sklearn.metrics
 
@@ -7,6 +8,19 @@ import sklearn.metrics
 def _check_sample_weight(sample_weight):
     if sample_weight is not None:
         raise ValueError("'sample_weight' is not supported.")
+
+
+def _check_reg_targets(y_true, y_pred, multioutput):
+    if multioutput != 'uniform_average':
+        raise NotImplementedError("'multioutput' must be 'uniform_average'")
+
+    if y_true.ndim == 1:
+        y_true = y_true.reshape((-1, 1))
+    if y_pred.ndim == 1:
+        y_pred = y_pred.reshape((-1, 1))
+
+    # TODO: y_type, multioutput
+    return None, y_true, y_pred, multioutput
 
 
 @doc_wraps(sklearn.metrics.mean_squared_error)
@@ -43,3 +57,26 @@ def mean_absolute_error(y_true, y_pred,
     else:
         raise ValueError("Weighted 'multioutput' not supported.")
     return output_errors.mean()
+
+
+@doc_wraps(sklearn.metrics.r2_score)
+def r2_score(y_true, y_pred, sample_weight=None,
+             multioutput="uniform_average"):
+    _check_sample_weight(sample_weight)
+    _, y_true, y_pred, multioutput = _check_reg_targets(
+        y_true, y_pred, multioutput
+    )
+    weight = 1.0
+
+    numerator = (weight * (y_true - y_pred) ** 2).sum(axis=0, dtype='f8')
+    denominator = (weight * (
+        y_true - y_true.mean(axis=0)) ** 2).sum(axis=0, dtype='f8')
+
+    nonzero_denominator = denominator != 0
+    nonzero_numerator = numerator != 0
+    valid_score = nonzero_denominator & nonzero_numerator
+    output_scores = da.ones([y_true.shape[1]], chunks=y_true.chunks[1])
+    output_scores[valid_score] = 1 - (numerator[valid_score] /
+                                      denominator[valid_score])
+    output_scores[nonzero_numerator & ~nonzero_denominator] = 0.
+    return output_scores.mean(axis=0)