+++ tests/metrics/test_regression.py
@@ -6,12 +6,15 @@ import sklearn.metrics
 import dask_ml.metrics
 
 
-@pytest.fixture(params=['mean_squared_error', 'mean_absolute_error'])
+@pytest.fixture(params=[
+    'mean_squared_error', 'mean_absolute_error', 'r2_score',
+])
 def metric_pairs(request):
     """Pairs of (dask-ml, sklearn) regression metrics.
 
     * mean_squared_error
     * mean_absolute_error
+    * r2_score
     """
     return (
         getattr(dask_ml.metrics, request.param),